Fix Resque and default backends
The interfaces were all out of sync.  Without a backend the purge command as concatenated twice leading to a "Wr 105
Unknown request." in varnish.  The ResqueJob was referencing VarnishInterface which isn't defined anywhere.  Although
the former was spec'ed, it was mocked out too aggressively and thus the problem was hidden.  For the job backends
I added new specs.gq