Fix telnet communication so Varnish#purge doesn't return false
Previously it was just processing the QUIT response.  Now it matches the
double newline as the prompt and returns the first line of the response.
If there is an error, additional info will be printed out, which will
currently be discarded.  Might be useful to store it somewhere, but my
main interest was just fixing the bug in #purge.